pyproject.toml
[project]
name = "python-ulid-django"
version = "1.2.0"
description = "ULID (Universally Unique Lexicographically Sortable Identifier) support for Django."
authors = [
    { name = "Xdynix", email = "Lizard.rar@gmail.com" },
]
dependencies = [
    "django>=4.2.0",
    "python-ulid<4.0.0,>=3.0.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = { text = "MIT" }
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Framework :: Django",
    "Framework :: Django :: 4",
    "Framework :: Django :: 4.1",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 5.1",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[project.urls]
Homepage = "https://github.com/Xdynix/python-ulid-django"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[dependency-groups]
lint = [
    "django-stubs[compatible-mypy]~=5.1.3",
    "mypy>=1.13.0",
    "pre-commit>=4.2.0",
]
test = [
    "beautifulsoup4[lxml]>=4.13.3",
    "pytest-cov>=6.0.0",
    "pytest-django>=4.10.0",
    "pytest-mock>=3.14.0",
    "pytest>=8.3.5",
]

[tool.pdm]
distribution = true

[tool.pdm.scripts.post_install]
cmd = ["pre-commit", "install"]
help = "install pre-commit hooks"

[tool.pdm.scripts.lint]
cmd = ["pre-commit", "run", "--all-files"]
env = { PYTHONUTF8 = "1" }
help = "execute linters"

[tool.pdm.scripts.test]
cmd = ["pytest", "--cov", "ulid_django"]
help = "execute test cases"

[tool.ruff]
cache-dir = ".cache/ruff"
target-version = "py310"
src = ["src", "testsite"]

[tool.ruff.lint]
select = [
    "F", # Pyflakes
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "I", # isort
    "PGH", # pygrep-hooks
    "UP", # pyupgrade
    "S", # flake8-bandit
    "B", # flake8-bugbear
    "C4", # flake8-comprehensions
    "DJ", # flake8-django
    "T20", # flake8-print
    "SIM", # flake8-simplify
    "ARG", # flake8-unused-arguments
    "PTH", # flake8-use-pathlib
    "FURB", # refurb
    "RUF", # Ruff-specific rules
]
ignore = [
    # Suppressed by ruff formatter.
    "W191", # tab-indentation
    "E111", # indentation-with-invalid-multiple
    "E114", # indentation-with-invalid-multiple-comment
    "E117", # over-indented
]

[tool.ruff.lint.per-file-ignores]
"test_*.py" = [
    "S101", # assert
]
"**/migrations/*.py" = [
    "RUF012", # mutable-class-default
]

[tool.mypy]
strict = true
cache_dir = ".cache/mypy"
mypy_path = "testsite"
plugins = [
    "mypy_django_plugin.main",
]

[tool.django-stubs]
django_settings_module = "testsite.settings"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "testsite.settings"
testpaths = "tests"
pythonpath = [".", "src", "testsite"]
cache_dir = ".cache/pytest"
addopts = [
    "--import-mode", "importlib",
    "--cov-config", "pyproject.toml",
    "--cov-report", "term",
    "--cov-report", "html",
    "--cov-fail-under", "100",
]

[tool.coverage.run]
branch = true
data_file = ".cache/coverage/data.db"
disable_warnings = ["no-data-collected"]

[tool.coverage.report]
exclude_also = [
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
]

[tool.coverage.html]
directory = ".cache/cov-report"
